# perf(overflow): Speed up overflow checks for small integers