--- a/go.mod
+++ b/go.mod
@@ -21,7 +21,6 @@ go 1.23.0
 toolchain go1.23.2
 
 require (
-	github.com/JohnCGriffin/overflow v0.0.0-20211019200055-46fa312c352c
 	github.com/andybalholm/brotli v1.1.1
 	github.com/apache/thrift v0.21.0
 	github.com/docopt/docopt-go v0.0.0-20180111231733-ee0de3bc6815