@cbandy 2025-03-07
It occurred to me today that this is better as two commits. The benchmarks are now committed, and one can try both implementations by switching between commits.

1. The first commit adds the new functions, their tests, and benchmarks for the current implementation.
2. The second commit uses the new functions throughout.