@zeroshade 2025-03-07
Thanks for this @cbandy! I'll take a closer look at it this afternoon!